+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.example.imagelisttest">
+    package="com.example.imageshow">
     <uses-permission android:name="android.permission.INTERNET"/>
 
     <application
@@ -9,15 +9,15 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.ImageListTest">
-        <activity android:name=".MainActivity">
+        android:theme="@style/Theme.ImageShow">
+        <activity android:name="com.example.imageshow.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".ui.PhotoDetailActivity"/>
+        <activity android:name="com.example.imageshow.ui.PhotoDetailActivity"/>
     </application>
 
 </manifest>
