--- app/src/main/java/com/example/imagelisttest/Config.kt
@@ -1,11 +1,0 @@
-package com.example.imagelisttest
-
-/**
- * Created by Ryu on 14,五月,2021
- */
-object Config {
-    const val API_KEY: String = "Mpm2euw4tMTyIWNVtdvLejuxn3IWaUV4G9j2NEu9ooA"
-    const val base_url = "https://api.unsplash.com/photos/?page=1&per_page=10&order_by=popular"
-    const val NETWORK_PAGE_SIZE = 50
-    const val PHOTO_DETAIL_TRANSITION_NAME = "detail:image"
-}
