--- app/src/main/java/com/example/imagelisttest/api/PhotoListResponse.kt
@@ -1,8 +1,0 @@
-package com.example.imagelisttest.api
-
-import com.example.imagelisttest.model.Photo
-
-/**
- * Created by Ryu on 16,五月,2021
- */
-class PhotoListResponse : ArrayList<Photo>()
