--- app/src/main/java/com/example/imagelisttest/model/Photo.kt
@@ -1,12 +1,0 @@
-package com.example.imagelisttest.model
-
-/**
- * Created by Ryu on 15,五月,2021
- * unsplash Urls
- */
-data class Photo(
-    val description: String,
-    val id: String,
-    val urls: Urls,
-    val user: User,
-)
