--- app/src/main/java/com/example/imagelisttest/model/Urls.kt
@@ -1,13 +1,0 @@
-package com.example.imagelisttest.model
-
-/**
- * Created by Ryu on 15,五月,2021
- * unsplash Urls
- */
-data class Urls(
-    val full: String,
-    val raw: String,
-    val regular: String,
-    val small: String,
-    val thumb: String
-)
