--- app/src/main/java/com/example/imagelisttest/model/User.kt
@@ -1,10 +1,0 @@
-package com.example.imagelisttest.model
-
-/**
- * Created by Ryu on 15,五月,2021
- * unsplash user
- */
-data class User(
-    val username: String,
-    val name: String,
-)
