--- app/src/main/java/com/example/imagelisttest/ui/adapter/OnItemClickListener.kt
@@ -1,11 +1,0 @@
-package com.example.imagelisttest.ui.adapter
-
-import androidx.appcompat.widget.AppCompatImageView
-import com.example.imagelisttest.model.Photo
-
-/**
- * Created by Ryu on 17,五月,2021
- */
-interface OnItemClickListener {
-    fun onItemClick(item: Photo?,view: AppCompatImageView)
-}
