+++ app/src/main/java/com/example/imageshow/api/PhotoListResponse.kt
@@ -1,0 +1,8 @@
+package com.example.imageshow.api
+
+import com.example.imageshow.model.Photo
+
+/**
+ * Created by Ryu on 16,五月,2021
+ */
+class PhotoListResponse : ArrayList<Photo>()
