+++ app/src/main/java/com/example/imageshow/model/Urls.kt
@@ -1,0 +1,13 @@
+package com.example.imageshow.model
+
+/**
+ * Created by Ryu on 15,五月,2021
+ * unsplash Urls
+ */
+data class Urls(
+    val full: String,
+    val raw: String,
+    val regular: String,
+    val small: String,
+    val thumb: String
+)
