+++ app/src/main/java/com/example/imageshow/model/User.kt
@@ -1,0 +1,10 @@
+package com.example.imageshow.model
+
+/**
+ * Created by Ryu on 15,五月,2021
+ * unsplash user
+ */
+data class User(
+    val username: String,
+    val name: String,
+)
