+++ app/src/main/java/com/example/imageshow/ui/PhotoFrameLayout.kt
@@ -1,0 +1,39 @@
+package com.example.imageshow.ui
+
+import android.content.Context
+import android.util.AttributeSet
+import android.widget.FrameLayout
+
+/**
+ * Created by Ryu on 17,五月,2021
+ */
+class PhotoFrameLayout constructor(context: Context, attrs: AttributeSet? = null) :
+    FrameLayout(context, attrs) {
+
+    override fun onMeasure(widthMeasureSpec: Int, heightMeasureSpec: Int) {
+        val widthSize = MeasureSpec.getSize(widthMeasureSpec)
+        val heightSize = MeasureSpec.getSize(heightMeasureSpec)
+
+        if (widthSize == 0 && heightSize == 0) {
+            // If there are no constraints on size, let FrameLayout measure
+            super.onMeasure(widthMeasureSpec, heightMeasureSpec)
+
+            val minSize = Math.min(measuredWidth, measuredHeight)
+            setMeasuredDimension(minSize, minSize)
+            return
+        }
+
+        val size = if (widthSize == 0 || heightSize == 0) {
+            // If one of the dimensions has no restriction on size, set both dimensions to be the
+            // on that does
+            Math.max(widthSize, heightSize)
+        } else {
+            // Both dimensions have restrictions on size, set both dimensions to be the
+            // smallest of the two
+            Math.min(widthSize, heightSize)
+        }
+
+        val newMeasureSpec = MeasureSpec.makeMeasureSpec(size, MeasureSpec.EXACTLY)
+        super.onMeasure(newMeasureSpec, newMeasureSpec)
+    }
+}
