+++ app/src/main/java/com/example/imageshow/ui/adapter/OnItemClickListener.kt
@@ -1,0 +1,11 @@
+package com.example.imageshow.ui.adapter
+
+import androidx.appcompat.widget.AppCompatImageView
+import com.example.imageshow.model.Photo
+
+/**
+ * Created by Ryu on 17,五月,2021
+ */
+interface OnItemClickListener {
+    fun onItemClick(item: Photo?, view: AppCompatImageView)
+}
