+++ app/src/test/java/com/example/imageshow/ui/PhotoListViewModelTest.kt
@@ -1,0 +1,27 @@
+package com.example.imageshow.ui
+
+import org.junit.After
+import org.junit.Before
+import org.junit.Test
+
+
+
+import org.junit.Assert.*
+
+/**
+ * Created by Ryu on 18,五月,2021
+ */
+class PhotoListViewModelTest {
+
+    @Before
+    fun setUp() {
+    }
+
+    @After
+    fun tearDown() {
+    }
+
+    @Test
+    fun getPhotoList() {
+    }
+}
