+++ build.gradle
@@ -31,17 +31,21 @@ ext{
     minSdkVersion = 15
     targetSdkVersion = 28
     supportLibVersion = '1.2.0'
-    coreVersion = '1.3.2'
+    coreVersion = '1.3.1'
     recyclerViewVersion = '1.2.0'
     constraintLayoutVersion = '2.0.4'
     materialVersion = '1.3.0'
-    lifecycleVersion = '2.3.1'
+    lifecycleVersion = '2.2.0'
     roomVersion = '2.3.0'
     pagingVersion = '3.0.0'
     retrofitVersion = '2.9.0'
+    gson = '2.9.0'
     okhttpLoggingInterceptorVersion = '4.9.0'
     coroutines = '1.4.3'
     junit = '4.+'
     androidxJunit = '1.1.2'
     espressoCore = '3.3.0'
+    mockwebserver = '4.7.2'
+    mockito = '3.+'
+    glide = '4.12.0'
 }
